Added nlohmann/json library for c++ interfaces.

+++ deploy/scripts/local_stop.sh
@@ -8,6 +8,7 @@
 INTERFACE_NAME=$1
 INTERFACE_DIRECTORY_PYTHON=RaspberryPi/interfaces/python
 INTERFACE_DIRECTORY_CPP=RaspberryPi/interfaces/c++
+LOCAL_INTERFACES_LIST=local_interfaces.conf
 
 if [ "${INTERFACE_NAME}" == "" ]; then
   echo "[Error] Please provide an interface name to launch or simply launch \"all\"."
@@ -18,31 +19,38 @@ fi
 if [ "${INTERFACE_NAME}" == "all" ]; then
   # Stop all interfaces if "all" is provided.
   STOPPED_SOMETHING=0
-  # All C++ interfaces.
-  for cpp_interface in $INTERFACE_DIRECTORY_CPP/*/; do
-    INAME=$(basename "$cpp_interface")
-    INTERFACE_PID=$(ps aux | grep c++/$INAME | grep -v "grep" | grep -v ".log" | awk '{print $2}' | head -n 1)
-    STATUS=$(kill $INTERFACE_PID 2>&1)
-    if [ "${STATUS}" == "" ]; then
-      echo ""
-      echo "Successfully stopped <$INAME>."
-      echo ""
-      STOPPED_SOMETHING=1
+  # If "all" is provided as the interface name, then all interfaces are
+  # stopped in the file local_interfaces.conf
+
+  # C++ interfaces
+  while IFS= read -r line; do
+    IPATH="${line// /}"
+    IFS='/' read -r ILANGUAGE INAME <<< "$IPATH"
+    if [[ "${IPATH}" != \#* ]]; then
+      if [[ "$ILANGUAGE" == "c++" ]]; then
+        INTERFACE_PID=$(ps aux | grep c++/$INAME | grep -v "grep" | grep -v ".log" | awk '{print $2}' | head -n 1)
+        STATUS=$(kill $INTERFACE_PID 2>&1)
+        if [ "${STATUS}" == "" ]; then
+          echo ""
+          echo "Successfully stopped <$INAME>."
+          echo ""
+          STOPPED_SOMETHING=1
+        fi
+      fi
+
+      # Python interfaces
+      if [[ "$ILANGUAGE" == "python" ]]; then
+        INTERFACE_PID=$(ps aux | grep python/$INAME | grep -v "grep" | grep -v ".log" | awk '{print $2}' | head -n 2)
+        STATUS=$(kill $INTERFACE_PID 2>&1)
+        if [ "${STATUS}" == "" ]; then
+          echo ""
+          echo "Successfully stopped <$INAME>."
+          echo ""
+          STOPPED_SOMETHING=1
+        fi
+      fi
     fi
-  done
-  
-  # All python interfaces.
-  for python_interface in $INTERFACE_DIRECTORY_PYTHON/*/; do
-    INAME=$(basename "$python_interface")
-    INTERFACE_PID=$(ps aux | grep python/$INAME | grep -v "grep" | grep -v ".log" | awk '{print $2}' | head -n 2)
-    STATUS=$(kill $INTERFACE_PID 2>&1)
-    if [ "${STATUS}" == "" ]; then
-      echo ""
-      echo "Successfully stopped <$INAME>."
-      echo ""
-      STOPPED_SOMETHING=1
-    fi
-  done
+  done < "$LOCAL_INTERFACES_LIST"
 
   if [ "$STOPPED_SOMETHING" == 0 ]; then
     echo ""
